update Recipe component state and add GetMe query
b: used prop drilling for state, as the component
used to be shallow. Wasn't fetching user at recipe
load, as it was unnecessary.

a: created RecipeContext, refeching query GetMe
on load, to be able to compare user with recipe author.
If user is the same as author, adding reviews will be
disabled.